requirements.txt
streamlit~=1.25.0
pandas~=2.0.2
matplotlib~=3.7.2
numpy~=1.25.0
pathlib~=1.0.1
scikit-learn~=1.2.2
lightgbm~=4.0.0
python-dateutil~=2.8.2